Solve #53 by using np.int64 instead of cython.int64

+++ Pyfhel/test.py
@@ -150,10 +150,10 @@ class PyfhelTestCase(unittest.TestCase):
         pyfhel.contextGen(p=1964769281, m=8192, base=2, sec=192, flagBatching=True)
         pyfhel.keyGen()
         self.assertTrue(pyfhel.batchEnabled())
-        ptxt = pyfhel.encodeArray(np.array([1, 2, 3, 4, 5, 6]))
+        ptxt = pyfhel.encodeArray(np.array([1, 2, 3, 4, 5, 6], dtype=np.int64))
         self.assertEqual(pyfhel.getnSlots(), 8192)
         self.assertTrue(
-            np.alltrue(pyfhel.decodeArray(ptxt)[:6] == np.array([1, 2, 3, 4, 5, 6]))
+            np.alltrue(pyfhel.decodeArray(ptxt)[:6] == np.array([1, 2, 3, 4, 5, 6], dtype=np.int64))
         )
 
     def test_Pyfhel_3_ENCRYPTING(self):
@@ -196,10 +196,10 @@ class PyfhelTestCase(unittest.TestCase):
         pyfhel.contextGen(p=1964769281, m=8192, base=2, sec=192, flagBatching=True)
         pyfhel.keyGen()
         self.assertTrue(pyfhel.batchEnabled())
-        ctxt = pyfhel.encryptArray(np.array([1, 2, 3, 4, 5, 6]))
+        ctxt = pyfhel.encryptArray(np.array([1, 2, 3, 4, 5, 6], dtype=np.int64))
         self.assertEqual(pyfhel.getnSlots(), 8192)
         self.assertTrue(
-            np.alltrue(pyfhel.decryptArray(ctxt)[:6] == np.array([1, 2, 3, 4, 5, 6]))
+            np.alltrue(pyfhel.decryptArray(ctxt)[:6] == np.array([1, 2, 3, 4, 5, 6], dtype=np.int64))
         )
 
     def test_Pyfhel_4_OPERATIONS(self):
@@ -268,8 +268,8 @@ class PyfhelTestCase(unittest.TestCase):
         pyfhel.keyGen()
         pyfhel.rotateKeyGen(60)
         ctxti = pyfhel.encryptBatch([1, 2, 3, 4, 5, 6])
-        ctxti2 = pyfhel.encryptArray(np.array([-6, -5, -4, -3, -2, -1]))
-        ptxti = pyfhel.encodeArray(np.array([12, 15, 18, 21, 24, 27]))
+        ctxti2 = pyfhel.encryptArray(np.array([-6, -5, -4, -3, -2, -1], dtype=np.int64))
+        ptxti = pyfhel.encodeArray(np.array([12, 15, 18, 21, 24, 27], dtype=np.int64))
 
         ctxt_add = pyfhel.add(ctxti, ctxti2, in_new_ctxt=True)
         ctxt_add2 = pyfhel.add_plain(ctxti, ptxti, in_new_ctxt=True)
